+++ ios/backgroundAudio/THEOplayerRCTView+BackgroundAudioConfig.swift
@@ -2,9 +2,11 @@
 
 import Foundation
 import THEOplayerSDK
+import AVFAudio
 
 struct BackgroundAudioConfig {
     var enabled: Bool = false
+    var shouldResumeAfterInterruption: Bool = false
 }
 
 extension THEOplayerRCTView: BackgroundPlaybackDelegate {
@@ -18,6 +20,9 @@ extension THEOplayerRCTView: BackgroundPlaybackDelegate {
             return
         }
         player.backgroundPlaybackDelegate = DefaultBackgroundPlaybackDelegate()
+        NotificationCenter.default.removeObserver(self,
+                                                  name: AVAudioSession.interruptionNotification,
+                                                  object: AVAudioSession.sharedInstance())
     }
     
     public func shouldContinueAudioPlaybackInBackground() -> Bool {
@@ -25,6 +30,52 @@ extension THEOplayerRCTView: BackgroundPlaybackDelegate {
         self.nowPlayingManager.updateNowPlaying()
         
         return self.backgroundAudioConfig.enabled
+    }
+    
+    func updateInterruptionNotifications() {
+        // Get the default notification center instance.
+        if self.backgroundAudioConfig.shouldResumeAfterInterruption {
+            NotificationCenter.default.addObserver(self,
+                                                   selector: #selector(handleInterruption),
+                                                   name: AVAudioSession.interruptionNotification,
+                                                   object: AVAudioSession.sharedInstance())
+        } else {
+            NotificationCenter.default.removeObserver(self,
+                                                      name: AVAudioSession.interruptionNotification,
+                                                      object: AVAudioSession.sharedInstance())
+        }
+    }
+    
+    
+    @objc func handleInterruption(notification: Notification) {
+        guard let userInfo = notification.userInfo,
+              let typeValue = userInfo[AVAudioSessionInterruptionTypeKey] as? UInt,
+              let type = AVAudioSession.InterruptionType(rawValue: typeValue) else {
+            return
+        }
+        
+        // Switch over the interruption type.
+        switch type {
+        case .began:
+            // An interruption began. Update the UI as necessary.
+            if DEBUG_INTERRUPTIONS { PrintUtils.printLog(logText: "[INTERRUPTION] An interruption began")}
+        case .ended:
+            // An interruption ended. Resume playback, if appropriate.
+            if DEBUG_INTERRUPTIONS { PrintUtils.printLog(logText: "[INTERRUPTION] An interruption ended")}
+            guard let optionsValue = userInfo[AVAudioSessionInterruptionOptionKey] as? UInt else { return }
+            let options = AVAudioSession.InterruptionOptions(rawValue: optionsValue)
+            if options.contains(.shouldResume) {
+                // An interruption ended. Resume playback.
+                if let player = self.player {
+                    if DEBUG_INTERRUPTIONS { PrintUtils.printLog(logText: "[INTERRUPTION] Ended interruption should resume playback => play()")}
+                    player.play()
+                }
+            } else {
+                // An interruption ended. Don't resume playback.
+                if DEBUG_INTERRUPTIONS { PrintUtils.printLog(logText: "[INTERRUPTION] Ended interruption should not resume playback.")}
+            }
+        default: ()
+        }
     }
 }