+++ src/api/backgroundAudio/BackgroundAudioConfiguration.ts
@@ -10,4 +10,12 @@ export interface BackgroundAudioConfiguration {
    * @defaultValue `false`
    */
   readonly enabled?: boolean;
+
+  /**
+   * Whether background audio should be resumed after an interruption (e.g. incoming call ended).
+   *
+   * @defaultValue `false`
+   * @remark Applies to iOS only, impacting behaviour for handling interruptions while on the lockscreen.
+   */
+  readonly shouldResumeAfterInterruption?: boolean;
 }